Removing Unused Code
This was originally a common library for simplifying support across MSVC or MinGW.  Currently nobody is known to be compiling our tests with MSVC, and I have little to no desire to support MSVC anyways, due to its generally inferior handling of C specific functions.  Clam's upstream repo still contains the source if it becomes useful at a later date.